id (from Int to Long)

+++ app/src/main/java/com/esmt/projet/victodo/feature_list/domain/model/TaskList.kt
@@ -2,11 +2,12 @@ package com.esmt.projet.victodo.feature_list.domain.model
 
 import androidx.compose.ui.graphics.Color
 import androidx.room.Entity
+import androidx.room.ForeignKey
 import androidx.room.PrimaryKey
 
 @Entity(tableName = "lists")
 data class TaskList(
-    @PrimaryKey val id: Int,
+    @PrimaryKey val id: Long,
     val title: String,
     val color: Int,//to check
     val isPinned: Boolean,
